split FindProp page components into own files, adjust Review & Send models
