+++ editor.planx.uk/src/@planx/components/FindProperty/model.ts
@@ -3,7 +3,9 @@ import { MoreInformation, parseMoreInformation } from "../shared";
 export interface FindProperty extends MoreInformation {
   title: string;
   description: string;
-  allowNewAddresses?: boolean;
+  allowNewAddresses: boolean;
+  newAddressTitle?: string;
+  newAddressDescription?: string;
 }
 
 export const parseFindProperty = (
@@ -12,36 +14,35 @@ export const parseFindProperty = (
   title: data?.title || "",
   description: data?.description || "",
   allowNewAddresses: data?.allowNewAddresses || false,
+  newAddressTitle: data?.newAddressTitle || "",
+  newAddressDescription: data?.newAddressDescription || "",
   ...parseMoreInformation(data),
 });
 
-export interface SiteAddress {
-  uprn: string;
-  blpu_code: string;
+// Minimum-required address details if proposing a new non-UPRN address
+export interface MinimumSiteAddress {
   latitude: number;
   longitude: number;
-  organisation: string | null;
-  sao: string | null;
-  pao: string;
-  street: string;
-  town: string;
-  postcode: string;
   x: number;
   y: number;
-  planx_description: string;
-  planx_value: string;
-  single_line_address: string;
   title: string;
-  source?: "Ordnance Survey AddressBase Premium";
+  source: "os" | "proposed";
 }
 
-export interface NewAddress {
-  latitude: number;
-  longitude: number;
-  x: number;
-  y: number;
-  title: string;
-  source?: "User input";
+// SiteAddress type reflects selecting a known address from the OS Places API "LPI" datasource
+export interface SiteAddress extends MinimumSiteAddress {
+  uprn?: string;
+  blpu_code?: string;
+  organisation?: string | null;
+  sao?: string | null;
+  pao?: string;
+  street?: string;
+  town?: string;
+  postcode?: string;
+  single_line_address?: string;
+  planx_description?: string; // joined via table blpu_codes
+  planx_value?: string; // joined via table blpu_codes
 }
 
 export const DEFAULT_TITLE = "Find the property" as const;
+export const DEFAULT_NEW_ADDRESS_TITLE = "Propose a new address" as const;
