+++ editor.planx.uk/src/@planx/components/shared/Preview/SummaryList.tsx
@@ -167,19 +167,36 @@ function Question(props: ComponentProps) {
 }
 
 function FindProperty(props: ComponentProps) {
-  const { postcode, single_line_address, town } = props.passport.data?._address;
-  return (
-    <>
-      <dt>Property</dt>
-      <dd>
-        {`${single_line_address.split(`, ${town}`)[0]}`}
-        <br />
-        {town}
-        <br />
-        {postcode}
-      </dd>
-    </>
-  );
+  const { source } = props.passport.data?._address;
+
+  if (source === "os") {
+    const { postcode, single_line_address, town } =
+      props.passport.data?._address;
+    return (
+      <>
+        <dt>Property</dt>
+        <dd>
+          {`${single_line_address.split(`, ${town}`)[0]}`}
+          <br />
+          {town}
+          <br />
+          {postcode}
+        </dd>
+      </>
+    );
+  } else {
+    const { x, y, title } = props.passport.data?._address;
+    return (
+      <>
+        <dt>Proposed address</dt>
+        <dd>
+          {`${title}`}
+          <br />
+          {`${x.toFixed(5)} Easting (X), ${y.toFixed(5)} Northing (Y)`}
+        </dd>
+      </>
+    );
+  }
 }
 
 function Checklist(props: ComponentProps) {
